backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pymupdf
python-multipart
supabase
requests
httpx